src/search/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
requests==2.31.0